+++ resources/views/sys/vm/item.blade.php
@@ -19,7 +19,8 @@
                     <tr class="d-none" vmSetting>
                         <td>parent</td>
                         <td class="text-end">
-                            <form class="d-block" method="post" action="{{ route('admin.vm.update', ['vm' => $id]) }}">
+                            <form class="d-block" method="post"
+                                action="{{ route('admin.vm.update', ['vm' => $vm_id]) }}">
                                 @method('PATCH') @csrf
                                 <select name="parent_id" onchange="this.form.submit()">
                                     <option value="" {{ $parent_id ?: 'selected' }}>none</option>
@@ -44,7 +45,7 @@
             <div class="card-footer bg-transparent">
                 <div class="row justify-content-between">
                     <div class="col-auto">
-                        <form class="d-block" method="post" action="{{ route('admin.vm.update', ['vm' => $id]) }}">
+                        <form class="d-block" method="post" action="{{ route('admin.vm.update', ['vm' => $vm_id]) }}">
                             @method('PATCH') @csrf
                             <input type="hidden" value="{{ !$is_allow }}" name="is_allow" />
                             <button type="submit" class="form-control form-control-sm btn text-danger">